+++ azplugins/__init__.py
@@ -7,3 +7,4 @@ import evaporate
 import pair
 import update
 import wall
+import restrain